requirements.txt
click
Flask
Flask-Cors
Flask-SQLAlchemy
greenlet
gunicorn
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
six
SQLAlchemy
zipp
psycopg2-binary
flask_httpauth
pillow
twilio
werkzeug
glob2

